:recycle:  refactor: 토큰 생성 및 파싱 방식 변경 (#16)
