+++ src/main/kotlin/com/kotlin/study/dongambackend/security/config/WebSecurityConfig.kt
@@ -22,10 +22,11 @@ class WebSecurityConfig(private val authenticationFilter: AuthenticationFilter) 
         .formLogin().disable()
         .headers { it.frameOptions().sameOrigin() }
         .authorizeHttpRequests {
-//            it.antMatchers("/api/user/sign-in", "/api/user/sign-up").permitAll()
+            it.antMatchers("/api/user/sign-in", "/api/user/sign-up").permitAll()
+                .antMatchers("/api/**").authenticated()
 //                .antMatchers("/api/admin/**").hasRole(UserRole.ADMIN.toString())
 //                .antMatchers("/api/**").hasRole(UserRole.USER.toString())
-            it.antMatchers("/api/**").permitAll()
+//            it.antMatchers("/api/**").permitAll()
         }
         .sessionManagement {
             it.sessionCreationPolicy(SessionCreationPolicy.STATELESS)
